refactor: helper function and potential badges name change
